pyproject.toml
[project]
name = "function-calling-demo"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0"
]
